requirements.txt
streamlit
pillow
python-dotenv
requests
openai
